requirements.txt
django
coverage
factory_boy
selenium
fabric
pyinotify
-e git://github.com/peterbe/python-gorun.git#egg=gorun
flake8
